refactor(dashboard): decouple monolithic DashboardPage into modular components
